requirements.txt
Flask==2.3.3
flask-cors==4.0.0
numpy==1.24.4
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
firebase-admin==6.4.0
google-cloud-storage==2.16.0
apify-client
gunicorn
mailersend
mercadopago==2.3.0
